refactor: split secret components into files
