+++ frontend/components/environments/SecretRow.tsx
@@ -1,798 +1,18 @@
-import {
-  ApiSecretEventEventTypeChoices,
-  EnvironmentType,
-  Maybe,
-  PersonalSecretType,
-  SecretTagType,
-  SecretType,
-} from '@/apollo/graphql'
-import { Fragment, useEffect, useState } from 'react'
-import {
-  FaEyeSlash,
-  FaEye,
-  FaTimes,
-  FaRegCommentDots,
-  FaTrashAlt,
-  FaHistory,
-  FaPlus,
-  FaTags,
-  FaCheckSquare,
-  FaSquare,
-  FaKey,
-  FaUserEdit,
-  FaTrash,
-  FaCheck,
-  FaDotCircle,
-  FaCircle,
-} from 'react-icons/fa'
+import { EnvironmentType, SecretType } from '@/apollo/graphql'
+import { useState } from 'react'
+import { FaEyeSlash, FaEye } from 'react-icons/fa'
 import { Button } from '../common/Button'
-import { Dialog, Switch, Transition } from '@headlessui/react'
-import { GetSecretTags } from '@/graphql/queries/secrets/getSecretTags.gql'
-import { CreateNewSecretTag } from '@/graphql/mutations/environments/createSecretTag.gql'
+
 import { LogSecretRead } from '@/graphql/mutations/environments/readSecret.gql'
-import { CreateNewPersonalSecret } from '@/graphql/mutations/environments/createPersonalSecret.gql'
-import { RemovePersonalSecret } from '@/graphql/mutations/environments/removePersonalSecret.gql'
 import clsx from 'clsx'
-import { relativeTimeFromDates } from '@/utils/time'
-import { useLazyQuery, useMutation } from '@apollo/client'
+import { useMutation } from '@apollo/client'
 import { areTagsAreSame } from '@/utils/tags'
-import { Avatar } from '../common/Avatar'
-import { SecretPropertyDiffs } from './SecretPropertyDiffs'
-import { encryptAsymmetric } from '@/utils/crypto'
-import { toast } from 'react-toastify'
 
-export const Tag = (props: { tag: SecretTagType }) => {
-  const { name, color } = props.tag
-
-  return (
-    <div className="flex items-center rounded-full gap-1 border border-zinc-300 dark:border-zinc-700 text-zinc-700 dark:text-zinc-200 text-base px-2">
-      <div className="flex items-center justify-center h-4 w-4">
-        <FaCircle className="text-sm" color={color} />
-      </div>
-      <span>{name}</span>
-    </div>
-  )
-}
-
-const TagsDialog = (props: {
-  orgId: string
-  secretId: string
-  secretName: string
-  tags: Array<SecretTagType>
-  handlePropertyChange: Function
-}) => {
-  const { orgId, secretId, secretName, tags, handlePropertyChange } = props
-
-  const [getOrgTags, { data: orgTags }] = useLazyQuery(GetSecretTags)
-  const [createSecretTag] = useMutation(CreateNewSecretTag)
-
-  const [secretTags, setSecretTags] = useState<Array<SecretTagType>>(tags)
-
-  const [isOpen, setIsOpen] = useState<boolean>(false)
-
-  const [newTag, setNewTag] = useState<Partial<SecretTagType>>({ name: '', color: '' })
-
-  useEffect(() => {
-    if (isOpen)
-      getOrgTags({
-        variables: {
-          orgId,
-        },
-      })
-  }, [getOrgTags, isOpen, orgId])
-
-  const closeModal = () => {
-    setIsOpen(false)
-  }
-
-  const openModal = () => {
-    setIsOpen(true)
-  }
-
-  const handleClose = () => {
-    handlePropertyChange(secretId, 'tags', secretTags)
-    closeModal()
-  }
-
-  const handleNewTagNameChange = (name: string) => setNewTag({ ...newTag, ...{ name } })
-
-  const handleNewTagColorChange = (color: string) => setNewTag({ ...newTag, ...{ color } })
-
-  const handleCreateTag = async () => {
-    const { name, color } = newTag
-
-    if (!name) return false
-
-    await createSecretTag({
-      variables: {
-        orgId,
-        name,
-        color,
-      },
-      refetchQueries: [
-        {
-          query: GetSecretTags,
-          variables: {
-            orgId,
-          },
-        },
-      ],
-    })
-    setNewTag({ name: '', color: '' })
-  }
-
-  const TagSelector = (props: { tag: SecretTagType }) => {
-    const { id, name, color } = props.tag
-
-    const isSelected = secretTags.map((secretTag) => secretTag.name).includes(name)
-
-    const handleTagClick = () => {
-      if (isSelected) {
-        setSecretTags(secretTags.filter((tag) => tag.name !== name))
-      } else setSecretTags([...secretTags, ...[{ id, name, color }]])
-    }
-
-    return (
-      <div
-        className="flex items-center gap-2 p-2 rounded-md cursor-pointer hover:bg-zinc-200 dark:hover:bg-zinc-600 transition-colors ease"
-        onClick={handleTagClick}
-      >
-        {isSelected ? (
-          <FaCheckSquare className="text-emerald-500" />
-        ) : (
-          <FaSquare className="text-zinc-300 dark:text-zinc-700" />
-        )}
-        <div className={clsx(isSelected ? 'opacity-100' : 'opacity-70', 'transition-opacity ease')}>
-          <Tag tag={props.tag} />
-        </div>
-      </div>
-    )
-  }
-
-  return (
-    <>
-      {tags.length > 0 ? (
-        <div className="flex items-center gap-1.5 cursor-pointer" role="button" onClick={openModal}>
-          {tags.map((tag) => (
-            <Tag key={tag.id} tag={tag} />
-          ))}
-        </div>
-      ) : (
-        <div className="flex items-center justify-center">
-          <Button variant="outline" onClick={openModal} title="Update tags" tabIndex={-1}>
-            <FaTags /> Tags
-          </Button>
-        </div>
-      )}
-
-      <Transition appear show={isOpen} as={Fragment}>
-        <Dialog as="div" className="relative z-10" onClose={closeModal}>
-          <Transition.Child
-            as={Fragment}
-            enter="ease-out duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in duration-200"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-black/25 backdrop-blur-md" />
-          </Transition.Child>
-
-          <div className="fixed inset-0 overflow-y-auto">
-            <div className="flex min-h-full items-center justify-center p-4 text-center">
-              <Transition.Child
-                as={Fragment}
-                enter="ease-out duration-300"
-                enterFrom="opacity-0 scale-95"
-                enterTo="opacity-100 scale-100"
-                leave="ease-in duration-200"
-                leaveFrom="opacity-100 scale-100"
-                leaveTo="opacity-0 scale-95"
-              >
-                <Dialog.Panel className="w-full max-w-2xl transform overflow-hidden rounded-2xl bg-neutral-100 dark:bg-neutral-900 p-6 text-left align-middle shadow-xl transition-all">
-                  <Dialog.Title as="div" className="flex w-full justify-between">
-                    <h3 className="text-lg font-medium leading-6 text-black dark:text-white ">
-                      Update{' '}
-                      <span className="text-zinc-700 dark:text-zinc-200 font-mono ph-no-capture">
-                        {secretName}
-                      </span>{' '}
-                      tags
-                    </h3>
-
-                    <Button variant="text" onClick={handleClose}>
-                      <FaTimes className="text-zinc-900 dark:text-zinc-400 hover:text-zinc-700 dark:hover:text-zinc-300" />
-                    </Button>
-                  </Dialog.Title>
-
-                  <div className="space-y-6 p-4">
-                    <div className="space-y-1">
-                      {orgTags?.secretTags.map((tag: SecretTagType) => (
-                        <TagSelector key={tag.id} tag={tag} />
-                      ))}
-                      <div className="flex items-center w-full justify-between border-t border-zinc-700 py-4">
-                        <div className="flex items-center gap-2">
-                          <input
-                            className="text-sm"
-                            value={newTag.name}
-                            onChange={(e) => handleNewTagNameChange(e.target.value)}
-                          />
-                          <input
-                            type="color"
-                            value={newTag.color}
-                            onChange={(e) => handleNewTagColorChange(e.target.value)}
-                          />
-                        </div>
-                        <Button
-                          variant={newTag.name!.length === 0 ? 'secondary' : 'primary'}
-                          disabled={newTag.name!.length === 0}
-                          onClick={handleCreateTag}
-                        >
-                          <FaPlus />
-                          Create new tag
-                        </Button>
-                      </div>
-                    </div>
-                  </div>
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </div>
-        </Dialog>
-      </Transition>
-    </>
-  )
-}
-
-const HistoryDialog = (props: { secret: SecretType; handlePropertyChange: Function }) => {
-  const { secret, handlePropertyChange } = props
-
-  const [isOpen, setIsOpen] = useState<boolean>(false)
-
-  const closeModal = () => {
-    setIsOpen(false)
-  }
-
-  const openModal = () => {
-    setIsOpen(true)
-  }
-
-  const getEventTypeColor = (eventType: ApiSecretEventEventTypeChoices) => {
-    if (eventType === ApiSecretEventEventTypeChoices.C) return 'bg-emerald-500'
-    if (eventType === ApiSecretEventEventTypeChoices.U) return 'bg-yellow-500'
-    if (eventType === ApiSecretEventEventTypeChoices.R) return 'bg-blue-500'
-    if (eventType === ApiSecretEventEventTypeChoices.D) return 'bg-red-500'
-  }
-
-  const getEventTypeText = (eventType: ApiSecretEventEventTypeChoices) => {
-    if (eventType === ApiSecretEventEventTypeChoices.C) return 'Created'
-    if (eventType === ApiSecretEventEventTypeChoices.U) return 'Updated'
-    if (eventType === ApiSecretEventEventTypeChoices.R) return 'Read'
-    if (eventType === ApiSecretEventEventTypeChoices.D) return 'Deleted'
-  }
-
-  const secretHistory = secret.history
-
-  return (
-    <>
-      <div className="flex items-center justify-center">
-        <Button variant="outline" onClick={openModal} title="View secret history" tabIndex={-1}>
-          <FaHistory /> <span className="hidden 2xl:block text-xs">History</span>
-        </Button>
-      </div>
-
-      <Transition appear show={isOpen} as={Fragment}>
-        <Dialog as="div" className="relative z-10" onClose={closeModal}>
-          <Transition.Child
-            as={Fragment}
-            enter="ease-out duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in duration-200"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-black/25 backdrop-blur-md" />
-          </Transition.Child>
-
-          <div className="fixed inset-0 overflow-y-auto">
-            <div className="flex min-h-full items-center justify-center p-4 text-center">
-              <Transition.Child
-                as={Fragment}
-                enter="ease-out duration-300"
-                enterFrom="opacity-0 scale-95"
-                enterTo="opacity-100 scale-100"
-                leave="ease-in duration-200"
-                leaveFrom="opacity-100 scale-100"
-                leaveTo="opacity-0 scale-95"
-              >
-                <Dialog.Panel className="w-full max-w-2xl transform overflow-hidden rounded-2xl bg-neutral-100 dark:bg-neutral-900 p-6 text-left align-middle shadow-xl transition-all">
-                  <Dialog.Title as="div" className="flex w-full justify-between items-start">
-                    <div>
-                      <h3 className="text-lg font-medium leading-6 text-black dark:text-white ">
-                        <span className="text-zinc-700 dark:text-zinc-200 font-mono ph-no-capture">
-                          {secret.key}
-                        </span>{' '}
-                        history
-                      </h3>
-                      <div className="text-neutral-500 text-sm">
-                        View the chronological history of changes made to this secret.
-                      </div>
-                    </div>
-
-                    <Button variant="text" onClick={closeModal}>
-                      <FaTimes className="text-zinc-900 dark:text-zinc-400 hover:text-zinc-700 dark:hover:text-zinc-300" />
-                    </Button>
-                  </Dialog.Title>
-
-                  <div className="space-y-8 py-4">
-                    <div className="max-h-[800px] overflow-y-auto px-2">
-                      <div className="space-y-4 pb-4 border-l border-zinc-300 dark:border-zinc-700">
-                        {secretHistory?.map((historyItem, index) => (
-                          <div key={historyItem!.timestamp} className="pb-8 space-y-2">
-                            <div className="flex flex-row items-center gap-2 -ml-1">
-                              <span
-                                className={clsx(
-                                  'h-2 w-2 rounded-full',
-                                  getEventTypeColor(historyItem!.eventType)
-                                )}
-                              ></span>
-                              <div className="text-zinc-800 dark:text-zinc-200 font-semibold">
-                                {getEventTypeText(historyItem!.eventType)}
-                              </div>
-                              <div className="text-neutral-500 text-sm">
-                                {relativeTimeFromDates(new Date(historyItem!.timestamp))}
-                              </div>{' '}
-                              <span className="text-sm text-neutral-500">by</span>
-                              <div className="text-sm flex items-center gap-2 text-neutral-500">
-                                {historyItem!.user ? (
-                                  <div className="flex items-center gap-1 text-sm">
-                                    <Avatar imagePath={historyItem!.user.avatarUrl!} size="sm" />
-                                    {historyItem?.user.fullName || historyItem?.user.email}
-                                  </div>
-                                ) : (
-                                  <div className="flex items-center gap-1 text-sm">
-                                    <FaKey /> Service token
-                                  </div>
-                                )}
-                              </div>
-                            </div>
-                            {index > 0 && (
-                              <SecretPropertyDiffs
-                                secret={secret}
-                                historyItem={historyItem!}
-                                index={index}
-                                handlePropertyChange={handlePropertyChange}
-                              />
-                            )}
-                          </div>
-                        ))}
-                      </div>
-                    </div>
-                  </div>
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </div>
-        </Dialog>
-      </Transition>
-    </>
-  )
-}
-
-const CommentDialog = (props: {
-  secretId: string
-  secretName: string
-  comment: string
-  handlePropertyChange: Function
-}) => {
-  const { secretId, secretName, comment, handlePropertyChange } = props
-
-  const [commentValue, setCommentValue] = useState<string>(comment)
-
-  const [isOpen, setIsOpen] = useState<boolean>(false)
-
-  const closeModal = () => {
-    setIsOpen(false)
-  }
-
-  const openModal = () => {
-    setIsOpen(true)
-  }
-
-  const handleClose = () => {
-    handlePropertyChange(secretId, 'comment', commentValue)
-    closeModal()
-  }
-
-  return (
-    <>
-      <div className="flex items-center justify-center">
-        <Button variant="outline" onClick={openModal} title="Update comment" tabIndex={-1}>
-          <FaRegCommentDots className={clsx(comment && 'text-emerald-500')} />{' '}
-          <span className="hidden 2xl:block text-xs">Comment</span>
-        </Button>
-      </div>
-
-      <Transition appear show={isOpen} as={Fragment}>
-        <Dialog as="div" className="relative z-10" onClose={handleClose}>
-          <Transition.Child
-            as={Fragment}
-            enter="ease-out duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in duration-200"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-black/25 backdrop-blur-md" />
-          </Transition.Child>
-
-          <div className="fixed inset-0 overflow-y-auto">
-            <div className="flex min-h-full items-center justify-center p-4 text-center">
-              <Transition.Child
-                as={Fragment}
-                enter="ease-out duration-300"
-                enterFrom="opacity-0 scale-95"
-                enterTo="opacity-100 scale-100"
-                leave="ease-in duration-200"
-                leaveFrom="opacity-100 scale-100"
-                leaveTo="opacity-0 scale-95"
-              >
-                <Dialog.Panel className="w-full max-w-2xl transform overflow-hidden rounded-2xl bg-neutral-100 dark:bg-neutral-900 p-6 text-left align-middle shadow-xl transition-all">
-                  <Dialog.Title as="div" className="flex w-full justify-between items-start">
-                    <div>
-                      <h3 className="text-lg font-medium leading-6 text-black dark:text-white ">
-                        Update{' '}
-                        <span className="text-zinc-700 dark:text-zinc-200 font-mono ph-no-capture">
-                          {secretName}
-                        </span>{' '}
-                        comment
-                      </h3>
-                      <div className="text-neutral-500 text-sm">
-                        Add a comment to this secret to provide additional information, context or
-                        instructions.
-                      </div>
-                    </div>
-
-                    <Button variant="text" onClick={handleClose}>
-                      <FaTimes className="text-zinc-900 dark:text-zinc-400 hover:text-zinc-700 dark:hover:text-zinc-300" />
-                    </Button>
-                  </Dialog.Title>
-
-                  <div className="space-y-6 p-4 ph-no-capture">
-                    <textarea
-                      rows={5}
-                      value={commentValue}
-                      className="w-full"
-                      onChange={(e) => setCommentValue(e.target.value)}
-                    ></textarea>
-                  </div>
-
-                  <div className="flex justify-end">
-                    <Button variant="secondary" onClick={handleClose}>
-                      Done
-                    </Button>
-                  </div>
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </div>
-        </Dialog>
-      </Transition>
-    </>
-  )
-}
-
-const OverrideDialog = (props: {
-  secretId: string
-  secretName: string
-  environment: EnvironmentType
-  override: Maybe<PersonalSecretType>
-}) => {
-  const { secretId, secretName, environment, override } = props
-
-  const [createOverride, { loading: createLoading }] = useMutation(CreateNewPersonalSecret)
-  const [removeOverride, { loading: removeLoading }] = useMutation(RemovePersonalSecret)
-
-  const [value, setValue] = useState<string>(override?.value || '')
-  const [isActive, setIsActive] = useState<boolean>(override ? override.isActive : true)
-
-  const [saved, setSaved] = useState<boolean>(true)
-  const [isOpen, setIsOpen] = useState<boolean>(false)
-
-  const reset = () => {
-    setValue(override?.value || '')
-    setIsActive(override?.isActive || true)
-  }
-
-  const clear = () => {
-    setValue('')
-    setIsActive(true)
-  }
-
-  const closeModal = () => {
-    setIsOpen(false)
-  }
-
-  const openModal = () => {
-    setIsOpen(true)
-  }
-
-  const valueUpdated = () => {
-    if (override === null && value === '') return false
-    else {
-      return override?.value !== value
-    }
-  }
-
-  const isActiveUpdated = () => {
-    if (override === null && value === '') return false
-    else {
-      return isActive !== override?.isActive
-    }
-  }
-
-  const saveRequired = valueUpdated() || isActiveUpdated()
-
-  useEffect(() => {
-    if (saveRequired) setSaved(false)
-  }, [saveRequired])
-
-  const toggleIsActive = () => setIsActive(!isActive)
-
-  const handleDeleteOverride = async () => {
-    await removeOverride({
-      variables: {
-        secretId,
-      },
-    })
-    clear()
-    toast.success('Removed personal secret')
-    closeModal()
-  }
-
-  const handleUpdateOverride = async () => {
-    const encryptedValue = await encryptAsymmetric(value, environment.identityKey)
-
-    await createOverride({
-      variables: {
-        newPersonalSecret: {
-          secretId,
-          value: encryptedValue,
-          isActive,
-        },
-      },
-    })
-    toast.success('Saved personal secret')
-    setSaved(true)
-    closeModal()
-  }
-
-  const handleClose = () => {
-    if (saveRequired && !saved) reset()
-    closeModal()
-  }
-
-  const activeOverride = override && override?.isActive
-
-  return (
-    <>
-      <div className="flex items-center justify-center">
-        <Button
-          variant="outline"
-          tabIndex={-1}
-          onClick={openModal}
-          title={
-            activeOverride ? 'A Personal Secret is overriding this value' : 'Override this value'
-          }
-        >
-          <FaUserEdit className={clsx(activeOverride && 'text-amber-500')} />{' '}
-          <span className={clsx('hidden 2xl:block text-xs', activeOverride && 'text-amber-500')}>
-            Override
-          </span>
-        </Button>
-      </div>
-
-      <Transition appear show={isOpen} as={Fragment}>
-        <Dialog as="div" className="relative z-10" onClose={handleClose}>
-          <Transition.Child
-            as={Fragment}
-            enter="ease-out duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in duration-200"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-black/25 backdrop-blur-md" />
-          </Transition.Child>
-
-          <div className="fixed inset-0 overflow-y-auto">
-            <div className="flex min-h-full items-center justify-center p-4 text-center">
-              <Transition.Child
-                as={Fragment}
-                enter="ease-out duration-300"
-                enterFrom="opacity-0 scale-95"
-                enterTo="opacity-100 scale-100"
-                leave="ease-in duration-200"
-                leaveFrom="opacity-100 scale-100"
-                leaveTo="opacity-0 scale-95"
-              >
-                <Dialog.Panel className="w-full max-w-2xl transform overflow-hidden rounded-2xl bg-neutral-100 dark:bg-neutral-900 p-6 text-left align-middle shadow-xl transition-all">
-                  <Dialog.Title as="div" className="flex w-full justify-between items-start">
-                    <div>
-                      <h3 className="text-lg font-medium leading-6 text-black dark:text-white ">
-                        Override{' '}
-                        <span className="text-zinc-700 dark:text-zinc-200 font-mono ph-no-capture">
-                          {secretName}
-                        </span>{' '}
-                      </h3>
-                      <div className="text-neutral-500 text-sm">
-                        Override this value with a Personal Secret. This value will only be visible
-                        to you, and will not alter the value of this secret for other users or 3rd
-                        party services.
-                      </div>
-                    </div>
-
-                    <Button variant="text" onClick={handleClose}>
-                      <FaTimes className="text-zinc-900 dark:text-zinc-400 hover:text-zinc-700 dark:hover:text-zinc-300" />
-                    </Button>
-                  </Dialog.Title>
-
-                  <div className="space-y-4 py-8">
-                    <div className="space-y-2">
-                      <div className="text-sm font-medium text-black dark:text-white">
-                        Override Value
-                      </div>
-                      <textarea
-                        rows={5}
-                        value={value}
-                        className="w-full ph-no-capture font-mono"
-                        onChange={(e) => setValue(e.target.value)}
-                      ></textarea>
-                    </div>
-                    <div className="flex items-center justify-start gap-2">
-                      <div className="text-sm text-black dark:text-white font-medium">
-                        Activate Secret Override
-                      </div>
-                      <Switch
-                        checked={isActive}
-                        onChange={toggleIsActive}
-                        className={`${
-                          isActive
-                            ? 'bg-emerald-400/10 ring-emerald-400/20'
-                            : 'bg-neutral-500/40 ring-neutral-500/30'
-                        } relative inline-flex h-6 w-11 items-center rounded-full ring-1 ring-inset`}
-                      >
-                        <span className="sr-only">Set as active</span>
-                        <span
-                          className={`${
-                            isActive ? 'translate-x-6 bg-emerald-400' : 'translate-x-1 bg-black'
-                          } flex items-center justify-center h-4 w-4 transform rounded-full transition`}
-                        ></span>
-                      </Switch>
-                    </div>
-                  </div>
-
-                  <div className="flex items-center justify-between gap-4">
-                    <div>
-                      {override && (
-                        <Button
-                          variant="danger"
-                          onClick={handleDeleteOverride}
-                          isLoading={removeLoading}
-                        >
-                          <FaTrash />
-                          Remove
-                        </Button>
-                      )}
-                    </div>
-                    <Button
-                      variant="primary"
-                      onClick={handleUpdateOverride}
-                      isLoading={createLoading}
-                      disabled={!saveRequired}
-                    >
-                      Save
-                    </Button>
-                  </div>
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </div>
-        </Dialog>
-      </Transition>
-    </>
-  )
-}
-
-const DeleteConfirmDialog = (props: {
-  secretId: string
-  secretName: string
-  onDelete: Function
-}) => {
-  const { secretName, secretId, onDelete } = props
-
-  const [isOpen, setIsOpen] = useState<boolean>(false)
-
-  const closeModal = () => {
-    setIsOpen(false)
-  }
-
-  const openModal = () => {
-    setIsOpen(true)
-  }
-
-  return (
-    <>
-      <div className="flex items-center justify-center">
-        <Button variant="danger" onClick={openModal} title="Delete secret" tabIndex={-1}>
-          <div className="text-white dark:text-red-500 flex items-center gap-1 p-1">
-            <FaTrashAlt />
-          </div>
-        </Button>
-      </div>
-
-      <Transition appear show={isOpen} as={Fragment}>
-        <Dialog as="div" className="relative z-10" onClose={closeModal}>
-          <Transition.Child
-            as={Fragment}
-            enter="ease-out duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in duration-200"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-black/25 backdrop-blur-md" />
-          </Transition.Child>
-
-          <div className="fixed inset-0 overflow-y-auto">
-            <div className="flex min-h-full items-center justify-center p-4 text-center">
-              <Transition.Child
-                as={Fragment}
-                enter="ease-out duration-300"
-                enterFrom="opacity-0 scale-95"
-                enterTo="opacity-100 scale-100"
-                leave="ease-in duration-200"
-                leaveFrom="opacity-100 scale-100"
-                leaveTo="opacity-0 scale-95"
-              >
-                <Dialog.Panel className="w-full max-w-2xl transform overflow-hidden rounded-2xl bg-neutral-100 dark:bg-neutral-900 p-6 text-left align-middle shadow-xl transition-all">
-                  <Dialog.Title as="div" className="flex w-full justify-between">
-                    <h3 className="text-lg font-medium leading-6 text-black dark:text-white ">
-                      Delete{' '}
-                      <span className="text-zinc-700 dark:text-zinc-200 font-mono ph-no-capture">
-                        {secretName}
-                      </span>
-                    </h3>
-
-                    <Button variant="text" onClick={closeModal}>
-                      <FaTimes className="text-zinc-900 dark:text-zinc-400 hover:text-zinc-700 dark:hover:text-zinc-300" />
-                    </Button>
-                  </Dialog.Title>
-
-                  <div className="space-y-6 p-4">
-                    <p className="text-neutral-500">Are you sure you want to delete this secret?</p>
-                    <div className="flex items-center gap-4">
-                      <Button variant="secondary" type="button" onClick={closeModal}>
-                        Cancel
-                      </Button>
-                      <Button variant="danger" onClick={() => onDelete(secretId)}>
-                        Delete
-                      </Button>
-                    </div>
-                  </div>
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </div>
-        </Dialog>
-      </Transition>
-    </>
-  )
-}
+import { DeleteConfirmDialog } from './secrets/DeleteDialog'
+import { CommentDialog } from './secrets/CommentDialog'
+import { HistoryDialog } from './secrets/HistoryDialog'
+import { OverrideDialog } from './secrets/OverrideDialog'
+import { TagsDialog } from './secrets/TagsDialog'
 
 export default function SecretRow(props: {
   orgId: string
