+++ frontend/components/environments/secrets/DeleteDialog.tsx
@@ -1,0 +1,91 @@
+import { Dialog, Transition } from '@headlessui/react'
+import { Button } from '../../common/Button'
+import { useState, Fragment } from 'react'
+import { FaTrashAlt, FaTimes } from 'react-icons/fa'
+
+export const DeleteConfirmDialog = (props: {
+  secretId: string
+  secretName: string
+  onDelete: Function
+}) => {
+  const { secretName, secretId, onDelete } = props
+
+  const [isOpen, setIsOpen] = useState<boolean>(false)
+
+  const closeModal = () => {
+    setIsOpen(false)
+  }
+
+  const openModal = () => {
+    setIsOpen(true)
+  }
+
+  return (
+    <>
+      <div className="flex items-center justify-center">
+        <Button variant="danger" onClick={openModal} title="Delete secret" tabIndex={-1}>
+          <div className="text-white dark:text-red-500 flex items-center gap-1 p-1">
+            <FaTrashAlt />
+          </div>
+        </Button>
+      </div>
+
+      <Transition appear show={isOpen} as={Fragment}>
+        <Dialog as="div" className="relative z-10" onClose={closeModal}>
+          <Transition.Child
+            as={Fragment}
+            enter="ease-out duration-300"
+            enterFrom="opacity-0"
+            enterTo="opacity-100"
+            leave="ease-in duration-200"
+            leaveFrom="opacity-100"
+            leaveTo="opacity-0"
+          >
+            <div className="fixed inset-0 bg-black/25 backdrop-blur-md" />
+          </Transition.Child>
+
+          <div className="fixed inset-0 overflow-y-auto">
+            <div className="flex min-h-full items-center justify-center p-4 text-center">
+              <Transition.Child
+                as={Fragment}
+                enter="ease-out duration-300"
+                enterFrom="opacity-0 scale-95"
+                enterTo="opacity-100 scale-100"
+                leave="ease-in duration-200"
+                leaveFrom="opacity-100 scale-100"
+                leaveTo="opacity-0 scale-95"
+              >
+                <Dialog.Panel className="w-full max-w-2xl transform overflow-hidden rounded-2xl bg-neutral-100 dark:bg-neutral-900 p-6 text-left align-middle shadow-xl transition-all">
+                  <Dialog.Title as="div" className="flex w-full justify-between">
+                    <h3 className="text-lg font-medium leading-6 text-black dark:text-white ">
+                      Delete{' '}
+                      <span className="text-zinc-700 dark:text-zinc-200 font-mono ph-no-capture">
+                        {secretName}
+                      </span>
+                    </h3>
+
+                    <Button variant="text" onClick={closeModal}>
+                      <FaTimes className="text-zinc-900 dark:text-zinc-400 hover:text-zinc-700 dark:hover:text-zinc-300" />
+                    </Button>
+                  </Dialog.Title>
+
+                  <div className="space-y-6 p-4">
+                    <p className="text-neutral-500">Are you sure you want to delete this secret?</p>
+                    <div className="flex items-center gap-4">
+                      <Button variant="secondary" type="button" onClick={closeModal}>
+                        Cancel
+                      </Button>
+                      <Button variant="danger" onClick={() => onDelete(secretId)}>
+                        Delete
+                      </Button>
+                    </div>
+                  </div>
+                </Dialog.Panel>
+              </Transition.Child>
+            </div>
+          </div>
+        </Dialog>
+      </Transition>
+    </>
+  )
+}
